Move report-error and robust to /lib/error
