+++ src/core/messenger.ts
@@ -1,4 +1,4 @@
-import { reportError } from '../utils/report-error'; // Keep this import relative otherwise Frontend will not be able to import this module. TODO: move all imports to relative imports
+import { reportError } from '../lib/error/report-error'; // Keep this import relative otherwise Frontend will not be able to import this module. TODO: move all imports to relative imports
 import { postMessage } from './messenger/post-message';
 
 /**
