+++ src/define/create-advert.ts
@@ -1,6 +1,6 @@
 import { log } from '@guardian/libs';
 import type { SizeMapping } from '../core/ad-sizes';
-import { reportError } from '../utils/report-error';
+import { reportError } from '../lib/error/report-error';
 import { Advert } from './Advert';
 
 const createAdvert = (
