+++ src/init/consented.ts
@@ -3,8 +3,8 @@ import { EventTimer } from '../core/event-timer';
 import { init as prepareAdVerification } from '../lib/ad-verification/prepare-ad-verification';
 import { commercialFeatures } from '../lib/commercial-features';
 import { adSlotIdPrefix } from '../lib/dfp/dfp-env-globals';
-import { reportError } from '../utils/report-error';
-import { catchErrorsAndReport } from '../utils/robust';
+import { reportError } from '../lib/error/report-error';
+import { catchErrorsAndReport } from '../lib/error/robust';
 import { adFreeSlotRemove } from './consented/ad-free-slot-remove';
 import { init as initComscore } from './consented/comscore';
 import { initDfpListeners } from './consented/dfp-listeners';
