+++ src/init/consented/dfp-listeners.ts
@@ -1,7 +1,7 @@
 import { onSlotLoad } from '../../events/on-slot-load';
 import { onSlotRender } from '../../events/on-slot-render';
 import { onSlotViewableFunction } from '../../events/on-slot-viewable';
-import { wrapWithErrorReporting } from '../../utils/report-error';
+import { wrapWithErrorReporting } from '../../lib/error/report-error';
 
 const initDfpListeners = (): Promise<void> => {
 	window.googletag.cmd.push(() => {
