+++ src/init/consented/dynamic-ad-slots.ts
@@ -5,7 +5,7 @@ import { init as initHighMerch } from '../../insert/high-merch';
 import { init as initMobileCrosswordsAdvert } from '../../insert/mobile-crossword-banner';
 import { init as initMobileSticky } from '../../insert/mobile-sticky';
 import { init as initLiveblogAdverts } from '../../insert/spacefinder/liveblog-adverts';
-import { reportError } from '../../utils/report-error';
+import { reportError } from '../../lib/error/report-error';
 import { initArticleBodyAdverts } from './article-body-adverts';
 
 type Modules = Array<[`${string}-${string}`, () => Promise<unknown>]>;
