+++ src/insert/fill-dynamic-advert-slot.ts
@@ -6,7 +6,7 @@ import { enableLazyLoad } from '../display/lazy-load';
 import { loadAdvert } from '../display/load-advert';
 import { dfpEnv } from '../lib/dfp/dfp-env';
 import { queueAdvert } from '../lib/dfp/queue-advert';
-import { reportError } from '../utils/report-error';
+import { reportError } from '../lib/error/report-error';
 
 const displayAd = (advert: Advert, forceDisplay: boolean) => {
 	if (dfpEnv.shouldLazyLoad() && !forceDisplay) {
