+++ src/lib/dfp/non-refreshable-line-items.ts
@@ -1,5 +1,5 @@
 import { memoize } from 'lodash-es';
-import { reportError } from '../../utils/report-error';
+import { reportError } from '../../lib/error/report-error';
 
 export const fetchNonRefreshableLineItemIds = async (): Promise<number[]> => {
 	// When the env is CODE or local, use the CODE env's non-refreshable line items file
