+++ src/lib/header-bidding/a9/a9.ts
@@ -1,7 +1,7 @@
 import { flatten } from 'lodash-es';
 import type { Advert } from '../../../define/Advert';
+import { reportError } from '../../../lib/error/report-error';
 import type { A9AdUnitInterface } from '../../../types/global';
-import { reportError } from '../../../utils/report-error';
 import type { HeaderBiddingSlot, SlotFlatMap } from '../prebid-types';
 import { getHeaderBiddingAdSlots } from '../slot-config';
